renamed DataBrowerFragment to DataBrowserFragment
